backend/recommendation/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
pika==1.2.0
PyJWT==2.3.0
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.27.1
